// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  CUSTOMER: 'CUSTOMER',
  COURIER: 'COURIER',
  ADMIN: 'ADMIN',
  SUPER_ADMIN: 'SUPER_ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const UserStatus = {
  ACTIVE: 'ACTIVE',
  BLOCKED: 'BLOCKED',
  DELETED: 'DELETED'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const AuthProvider = {
  CREDENTIALS: 'CREDENTIALS',
  GOOGLE: 'GOOGLE'
} as const

export type AuthProvider = (typeof AuthProvider)[keyof typeof AuthProvider]


export const ParcelStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  ASSIGNED: 'ASSIGNED',
  PICKED_UP: 'PICKED_UP',
  IN_TRANSIT: 'IN_TRANSIT',
  DELIVERED: 'DELIVERED',
  CANCELLED: 'CANCELLED',
  FAILED: 'FAILED'
} as const

export type ParcelStatus = (typeof ParcelStatus)[keyof typeof ParcelStatus]


export const ParcelType = {
  DOCUMENT: 'DOCUMENT',
  PACKAGE: 'PACKAGE',
  FRAGILE: 'FRAGILE',
  ELECTRONICS: 'ELECTRONICS',
  OTHER: 'OTHER'
} as const

export type ParcelType = (typeof ParcelType)[keyof typeof ParcelType]


export const PaymentStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID',
  FAILED: 'FAILED',
  REFUNDED: 'REFUNDED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const PaymentGateway = {
  STRIPE: 'STRIPE',
  BKASH: 'BKASH'
} as const

export type PaymentGateway = (typeof PaymentGateway)[keyof typeof PaymentGateway]